move editor into source
