--- plugins/editor/styles.js
@@ -1,14 +1,0 @@
-'use strict';
-
-const styles = {
-  baud: {
-    float: 'left',
-    padding: '3px 0px 0px 10px',
-    margin: 0
-  },
-  baudRate: {
-    paddingLeft: 8
-  }
-};
-
-module.exports = styles;
